Code cleanup.

+++ src/main/java/org/halvors/nuclearphysics/api/effect/explosion/IFulmination.java
@@ -2,9 +2,9 @@ package org.halvors.nuclearphysics.api.effect.explosion;
 
 public interface IFulmination {
     /**
-     * The RADIUS of effect of the explosion.
+     * The radius of effect of the explosion.
      */
-    float getRadius();
+    double getRadius();
 
     /**
      * The energy emitted by this explosive. In Joules and approximately based off of a real life equivalent.
