+++ src/main/java/org/halvors/nuclearphysics/common/effect/explosion/AntimatterExplosion.java
@@ -25,7 +25,7 @@ public class AntimatterExplosion extends RadioactiveExplosion implements IFulmin
     }
 
     @Override
-    public float getRadius() {
+    public double getRadius() {
         return size;
     }
